src/mfaktc.c:
Assing 0 to errno at the start of the loop that parses the command
line; otherwise, valid command lines fail because something before
that loops sets errno on Cygwin/MSWindows10, at least for me.

Also, change the type of exponent, bit_min, and bit_max to match the
return values of strtoul() and strtol() and drop the type casting that
is no longer needed.