Fix barrett15 with VectorSize=1
The definition of SUB_COND was passing non-boolean values to AS_UINT_V, which
was meant only for results of boolean operations. As a result, SUB_COND would
increment rather than decrement the result for the true condition for
VectorSize=1. SUB_COND is used in barrett15.cl.

Convert AS_UINT_V to an inline function (it should be renamed later). Ensure
that AS_UINT_V only takes bool (it can only be done for VectorSize=1, otherwise
it takes int_v). Remove AS_INT_V, AS_LONG_V and AS_ULONG_V, they are unused and
it's not clear if they were meant for booleans only.

Remove workaround for OpenCL implementations using 1 for true in vectors. The
requirement to use -1 for true in vectors goes back to at least OpenCL 1.1.

There is no way mfakto would have worked for any implementation that is still
using 1 for true in vectors. The workaround was only applied to ADD_COND and
SUB_COND but not to AS_UINT_V, which is used all over the place in a way that
would break for non-compliant OpenCL implementations.